[ReignOfKings] Update to Visual Studio 2017 project format
