+++ Games/Oxide.ReignOfKings/Files/_start-example.bat
@@ -1,0 +1,12 @@
+@echo off
+cls
+:start
+echo Starting server...
+
+ROK.exe -batchmode -nographics -silentcrash
+
+echo.
+echo Restarting server...
+timeout /t 10
+echo.
+goto start
