+++ Games/Oxide.ReignOfKings/Libraries/Covalence/ReignOfKingsPlayer.cs
@@ -1,0 +1,364 @@
+﻿using System;
+using System.Globalization;
+using CodeHatch.Common;
+using CodeHatch.Damaging;
+using CodeHatch.Engine.Behaviours;
+using CodeHatch.Engine.Core.Commands;
+using CodeHatch.Engine.Networking;
+using CodeHatch.Networking.Events;
+using CodeHatch.Networking.Events.Entities;
+using CodeHatch.StarForge.Sleeping;
+using Oxide.Core;
+using Oxide.Core.Libraries;
+using Oxide.Core.Libraries.Covalence;
+using UnityEngine;
+
+namespace Oxide.Game.ReignOfKings.Libraries.Covalence
+{
+    /// <summary>
+    /// Represents a player, either connected or not
+    /// </summary>
+    public class ReignOfKingsPlayer : IPlayer, IEquatable<IPlayer>
+    {
+        private static Permission libPerms;
+        private readonly Player player;
+        private readonly ulong steamId;
+
+        internal ReignOfKingsPlayer(ulong id, string name)
+        {
+            if (libPerms == null) libPerms = Interface.Oxide.GetLibrary<Permission>();
+
+            Name = name;
+            steamId = id;
+            Id = id.ToString();
+        }
+
+        internal ReignOfKingsPlayer(Player player) : this(player.Id, player.Name)
+        {
+            this.player = player;
+        }
+
+        #region Objects
+
+        /// <summary>
+        /// Gets the object that backs the player
+        /// </summary>
+        public object Object => player;
+
+        /// <summary>
+        /// Gets the player's last command type
+        /// </summary>
+        public CommandType LastCommand { get; set; }
+
+        #endregion
+
+        #region Information
+
+        /// <summary>
+        /// Gets the name for the player
+        /// </summary>
+        public string Name { get; }
+
+        /// <summary>
+        /// Gets the ID for the player (unique within the current game)
+        /// </summary>
+        public string Id { get; }
+
+        /// <summary>
+        /// Gets the player's language
+        /// </summary>
+        public CultureInfo Language => CultureInfo.GetCultureInfo("en"); // TODO: Implement when possible
+
+        /// <summary>
+        /// Gets the player's IP address
+        /// </summary>
+        public string Address => player.Connection.IpAddress;
+
+        /// <summary>
+        /// Gets the player's average network ping
+        /// </summary>
+        public int Ping => player.Connection.AveragePing;
+
+        /// <summary>
+        /// Returns if the player is admin
+        /// </summary>
+        public bool IsAdmin => player?.HasPermission("admin") ?? false;
+
+        /// <summary>
+        /// Gets if the player is banned
+        /// </summary>
+        public bool IsBanned => Server.IdIsBanned(steamId);
+
+        /// <summary>
+        /// Gets if the player is connected
+        /// </summary>
+        public bool IsConnected => player?.Connection?.IsConnected ?? false;
+
+        /// <summary>
+        /// Returns if the player is sleeping
+        /// </summary>
+        public bool IsSleeping
+        {
+            get
+            {
+                var sleeper = player?.Entity.Get<ISleeper>();
+                return sleeper != null && sleeper.IsSleeping;
+            }
+        }
+
+        /// <summary>
+        /// Returns if the player is the server
+        /// </summary>
+        public bool IsServer => false;
+
+        #endregion
+
+        #region Administration
+
+        /// <summary>
+        /// Bans the player for the specified reason and duration
+        /// </summary>
+        /// <param name="reason"></param>
+        /// <param name="duration"></param>
+        public void Ban(string reason, TimeSpan duration = default(TimeSpan))
+        {
+            // Check if already banned
+            if (IsBanned) return;
+
+            // Ban and kick user
+            Server.Ban(steamId, (int)duration.TotalSeconds, reason);
+        }
+
+        /// <summary>
+        /// Gets the amount of time remaining on the player's ban
+        /// </summary>
+        public TimeSpan BanTimeRemaining => new DateTime(Server.GetBannedPlayerFromPlayerId(steamId).ExpireDate) - DateTime.Now;
+
+        /// <summary>
+        /// Heals the player's character by specified amount
+        /// </summary>
+        /// <param name="amount"></param>
+        public void Heal(float amount) => player.Heal(amount);
+
+        /// <summary>
+        /// Gets/sets the player's health
+        /// </summary>
+        public float Health
+        {
+            get { return player.GetHealth().CurrentHealth; }
+            set { player.GetHealth().CurrentHealth = value; }
+        }
+
+        /// <summary>
+        /// Damages the player's character by specified amount
+        /// </summary>
+        /// <param name="amount"></param>
+        public void Hurt(float amount)
+        {
+            var damage = new Damage(player.CurrentCharacter.Prefab, null)
+            {
+                Amount = amount,
+                DamageTypes = DamageType.Unknown,
+                Damager = player.Entity
+            };
+            EventManager.CallEvent(new EntityDamageEvent(player.Entity, damage));
+        }
+
+        /// <summary>
+        /// Kicks the player from the game
+        /// </summary>
+        /// <param name="reason"></param>
+        public void Kick(string reason) => Server.Kick(player, reason);
+
+        /// <summary>
+        /// Causes the player's character to die
+        /// </summary>
+        public void Kill() => player.Kill();
+
+        /// <summary>
+        /// Gets/sets the player's maximum health
+        /// </summary>
+        public float MaxHealth
+        {
+            get { return player.GetHealth().MaxHealth; }
+            set { player.GetHealth().MaxHealth = value; }
+        }
+
+        /// <summary>
+        /// Renames the player to specified name
+        /// <param name="name"></param>
+        /// </summary>
+        public void Rename(string name)
+        {
+            // TODO: Implement when possible
+        }
+
+        /// <summary>
+        /// Teleports the player's character to the specified position
+        /// </summary>
+        /// <param name="x"></param>
+        /// <param name="y"></param>
+        /// <param name="z"></param>
+        public void Teleport(float x, float y, float z) => player.Entity.GetOrCreate<CharacterTeleport>().Teleport(new Vector3(x, y, z));
+
+        /// <summary>
+        /// Teleports the player's character to the specified generic position
+        /// </summary>
+        /// <param name="pos"></param>
+        public void Teleport(GenericPosition pos) => Teleport(pos.X, pos.Y, pos.Z);
+
+        /// <summary>
+        /// Unbans the player
+        /// </summary>
+        public void Unban()
+        {
+            // Check if unbanned already
+            if (!IsBanned) return;
+
+            // Set to unbanned
+            Server.Unban(steamId);
+        }
+
+        #endregion
+
+        #region Location
+
+        /// <summary>
+        /// Gets the position of the player
+        /// </summary>
+        /// <param name="x"></param>
+        /// <param name="y"></param>
+        /// <param name="z"></param>
+        public void Position(out float x, out float y, out float z)
+        {
+            var pos = player.CurrentCharacter.SavedPosition;
+            x = pos.x;
+            y = pos.y;
+            z = pos.z;
+        }
+
+        /// <summary>
+        /// Gets the position of the player
+        /// </summary>
+        /// <returns></returns>
+        public GenericPosition Position()
+        {
+            var pos = player.CurrentCharacter.SavedPosition;
+            return new GenericPosition(pos.x, pos.y, pos.z);
+        }
+
+        #endregion
+
+        #region Chat and Commands
+
+        /// <summary>
+        /// Sends the specified message to the player
+        /// </summary>
+        /// <param name="message"></param>
+        public void Message(string message) => player.SendMessage(message);
+
+        /// <summary>
+        /// Sends the specified message to the player
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="args"></param>
+        public void Message(string message, params object[] args) => Message(string.Format(message, args));
+
+        /// <summary>
+        /// Replies to the player with the specified message
+        /// </summary>
+        /// <param name="message"></param>
+        public void Reply(string message) => Message(message);
+
+        /// <summary>
+        /// Replies to the player with the specified message
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="args"></param>
+        public void Reply(string message, params object[] args) => Message(message, args);
+
+        /// <summary>
+        /// Runs the specified console command on the player
+        /// </summary>
+        /// <param name="command"></param>
+        /// <param name="args"></param>
+        public void Command(string command, params object[] args)
+        {
+            CommandManager.ExecuteCommand(steamId, $"{command} {string.Join(" ", Array.ConvertAll(args, x => x.ToString()))}");
+        }
+
+        #endregion
+
+        #region Permissions
+
+        /// <summary>
+        /// Gets if the player has the specified permission
+        /// </summary>
+        /// <param name="perm"></param>
+        /// <returns></returns>
+        public bool HasPermission(string perm) => libPerms.UserHasPermission(Id, perm);
+
+        /// <summary>
+        /// Grants the specified permission on this player
+        /// </summary>
+        /// <param name="perm"></param>
+        public void GrantPermission(string perm) => libPerms.GrantUserPermission(Id, perm, null);
+
+        /// <summary>
+        /// Strips the specified permission from this player
+        /// </summary>
+        /// <param name="perm"></param>
+        public void RevokePermission(string perm) => libPerms.RevokeUserPermission(Id, perm);
+
+        /// <summary>
+        /// Gets if the player belongs to the specified group
+        /// </summary>
+        /// <param name="group"></param>
+        /// <returns></returns>
+        public bool BelongsToGroup(string group) => libPerms.UserHasGroup(Id, group);
+
+        /// <summary>
+        /// Adds the player to the specified group
+        /// </summary>
+        /// <param name="group"></param>
+        public void AddToGroup(string group) => libPerms.AddUserGroup(Id, group);
+
+        /// <summary>
+        /// Removes the player from the specified group
+        /// </summary>
+        /// <param name="group"></param>
+        public void RemoveFromGroup(string group) => libPerms.RemoveUserGroup(Id, group);
+
+        #endregion
+
+        #region Operator Overloads
+
+        /// <summary>
+        /// Returns if player's unique ID is equal to another player's unique ID
+        /// </summary>
+        /// <param name="other"></param>
+        /// <returns></returns>
+        public bool Equals(IPlayer other) => Id == other?.Id;
+
+        /// <summary>
+        /// Returns if player's object is equal to another player's object
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <returns></returns>
+        public override bool Equals(object obj) => obj is IPlayer && Id == ((IPlayer)obj).Id;
+
+        /// <summary>
+        /// Gets the hash code of the player's unique ID
+        /// </summary>
+        /// <returns></returns>
+        public override int GetHashCode() => Id.GetHashCode();
+
+        /// <summary>
+        /// Returns a human readable string representation of this IPlayer
+        /// </summary>
+        /// <returns></returns>
+        public override string ToString() => $"Covalence.ReignOfKingsPlayer[{Id}, {Name}]";
+
+        #endregion
+    }
+}
