+++ Games/Oxide.ReignOfKings/ReignOfKingsPluginLoader.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Oxide.Core.Plugins;
+
+namespace Oxide.Game.ReignOfKings
+{
+    /// <summary>
+    /// Responsible for loading Reign of Kings core plugins
+    /// </summary>
+    public class ReignOfKingsPluginLoader : PluginLoader
+    {
+        public override Type[] CorePlugins => new[] { typeof(ReignOfKingsCore) };
+    }
+}
