--- Games/Unity/Oxide.Game.ReignOfKings/Files/_start-example.bat
@@ -1,12 +1,0 @@
-@echo off
-cls
-:start
-echo Starting server...
-
-ROK.exe -batchmode -nographics -silentcrash
-
-echo.
-echo Restarting server...
-timeout /t 10
-echo.
-goto start
