+++ Oxide.sln
@@ -90,7 +90,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlanetExplorers", "Games\Un
 		{52C26C5B-23E5-48A2-B403-5FF5C898E91A} = {52C26C5B-23E5-48A2-B403-5FF5C898E91A}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReignOfKings", "Games\Unity\Oxide.Game.ReignOfKings\ReignOfKings.csproj", "{E6788452-90F2-4388-A9BC-67D7EBBFF8CD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReignOfKings", "Games\Oxide.ReignOfKings\ReignOfKings.csproj", "{E6788452-90F2-4388-A9BC-67D7EBBFF8CD}"
 	ProjectSection(ProjectDependencies) = postProject
 		{C54EA613-C5FD-4ACD-8AC4-79063B08CF5A} = {C54EA613-C5FD-4ACD-8AC4-79063B08CF5A}
 		{D138493B-6F2E-4B9F-8F79-17B73422D87F} = {D138493B-6F2E-4B9F-8F79-17B73422D87F}
